// client/dist/widgets/lrs/search-by-route/src/config.ts
export enum Identifiers {
  RouteId = 'ROUTEID',
  RouteName = 'ROUTENAME',
  MultiField = 'MULTIFIELD'
}

export enum LineIdentifiers {
  LineId = 'LINEID',
  LineName = 'LINENAME'
}

export enum SpatialReferenceFrom {
  Map = 'MAP',
  Lrs = 'LRS'
}
// ...
export interface RouteIdFieldsSettings {
  field?: IMFieldSchema
  enabled?: boolean
}

export interface SpatialReferenceInfo {
  wkid: number
  wkt: string
}

export interface NetworkItem {
  id?: string
  name?: string
  configId?: string
  layerId?: string
  lrsNetworkId?: string
  outputLineDsId?: string
  outputPointDsId?: string
  networkUrl: string
  useRouteId?: boolean
  useRouteName?: boolean
  useLineName?: boolean
  useLineId?: boolean
  useMultiField?: boolean
  useMeasure?: boolean
  useCoordinate?: boolean
  useReferent?: boolean
  useLineAndMeasure?: boolean
  useDataSource?: UseDataSource
  defaultMethod?: SearchMethod
  enabledMethods?: SearchMethod[]
  defaultIdentifer?: Identifiers
  defaultLineIdentifier?: LineIdentifiers
  routeIdFieldName?: string
  routeNameFieldName?: string
  routeIdFieldNames?: string[]
  routeIdField?: IMFieldSchema
  routeNameField?: IMFieldSchema
  lineIdField?: IMFieldSchema
  lineNameField?: IMFieldSchema
  lineOrderField?: IMFieldSchema
  fromDateField?: IMFieldSchema
  toDateField?: IMFieldSchema
  lineIdFieldName?: string
  lineNameFieldName?: string
  lineOrderFieldName?: string
  routeIdFields?: RouteIdFieldsSettings[]
  supportsLines?: boolean
  showLineAndMeasure?: boolean
  unitsOfMeasure?: string
  sortOptions?: OrderByOption[]
  resultDisplayFields?: string[]
  measurePrecision: number
  defaultSpatialReferenceFrom?: SpatialReferenceFrom
  searchRadius?: number // in network's units of measure
  referent?: boolean
  type?: string
  layerFields?: any
  displayName?: string
  fromDateFieldName?: string
  toDateFieldName?: string
  searchSingle?: boolean
  searchMultiple?: boolean
  searchRange?: boolean
  expandByDefault?: boolean
  spatialReferenceInfo?: SpatialReferenceInfo
  derivedFromNetworkId: string
  isDerived: boolean
  lrsFields?: string[]
  additionalFields?: string[]
  defaultChecked?: string[]
  showAdditionalFields?: boolean
  allAdditionalFieldsChecked?: boolean
  allNetworkFieldsChecked?: boolean
  allFieldsDetails?: IMFieldSchema[]
}